Rename dependencies.py back to auth.py
