+++ api/app/auth.py
@@ -1,5 +1,4 @@
-""" Authentication module for validating access tokens generated by Keycloak
-"""
+"""Dependency functions used for authenticating and auditing requests"""
 import logging
 from fastapi import Depends, HTTPException, status, Request
 from fastapi.security import OAuth2PasswordBearer
