+++ api/app/main.py
@@ -8,7 +8,7 @@ from fastapi.middleware.cors import CORSMiddleware
 from starlette.applications import Starlette
 from app import schemas, configure_logging
 from app.percentile import get_precalculated_percentiles
-from app.dependency import authentication_required, audit
+from app.auth import authentication_required, audit
 from app import config
 from app import health
 from app import hourlies
